requirements.txt
fastapi
uvicorn
langchain
langchain_community
langchain_openai
python_dotenv
pydantic
openai
tiktoken
SQLAlchemy
